% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton.R
\name{generate_skeleton}
\alias{generate_skeleton}
\title{Add together two numbers}
\usage{
generate_skeleton(...)
}
\arguments{
\item{...}{This is a wrapper function, see details}
}
\value{
A number.
}
\description{
Add together two numbers
}
\details{
Coming soon
COming soon
}
\examples{
add(1, 1)
add(10, 1)
}
