% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton_defaults.R
\name{generate_skeleton_defaults}
\alias{generate_skeleton_defaults}
\title{Add together two numbers}
\usage{
generate_skeleton_defaults(
  packages_yaml_file_path = system.file("defaults/", package = "PkgFamilySite"),
  path
)
}
\arguments{
\item{packages_yaml_file_path}{Path to the packages.yaml file}

\item{path}{Path to where the site is. Mus end in /}
}
\value{
A number.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
