% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton_index.R
\name{generate_skeleton_index}
\alias{generate_skeleton_index}
\title{Add together two numbers}
\usage{
generate_skeleton_index(family_name, family_description, path)
}
\arguments{
\item{path}{Path to where the site is. Mus end in /}

\item{x}{A number.}
}
\value{
A number.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
