% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton_logo.R
\name{generate_skeleton_logo}
\alias{generate_skeleton_logo}
\title{Add together two numbers}
\usage{
generate_skeleton_logo(logo_url, path)
}
\arguments{
\item{logo_url}{Url of the .png logo file of the form \code{"http://logo.com/logo.png"}}

\item{path}{Path to where the site is. Mus end in /}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
