% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_skeleton_quartoyaml.R
\name{generate_skeleton_quartoyaml}
\alias{generate_skeleton_quartoyaml}
\title{Add together two numbers}
\usage{
generate_skeleton_quartoyaml(
  family_name,
  family_description,
  family_site_url = "/",
  github_link = "",
  google_analytics = "",
  path
)
}
\arguments{
\item{family_name}{The name of this family of packages.}

\item{family_description}{Description of this family of packages.}

\item{family_site_url}{What link will this site be on.}

\item{github_link}{What is the link to the github repo (or github organisation)}

\item{google_analytics}{GA-XXXXXX number.}

\item{path}{Path to where the site is. Mus end in /}
}
\value{
A number.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
